Extracted strike actions from TraineeController to a new StrikeController
