+++ server/src/controllers/index.ts
@@ -4,6 +4,7 @@ export * from './Trainee/TraineeController';
 export * from './Trainee/InteractionController';
 export * from './Trainee/TestController';
 export * from './Trainee/ProfilePictureController';
+export * from './Trainee/StrikeController';
 export * from './GeographyController';
 export * from './DashboardController';
 export * from './CohortsController';
