+++ server/src/index.ts
@@ -23,6 +23,7 @@ import {
   CohortsController,
   TestController,
   ProfilePictureController,
+  StrikeController,
 } from './controllers';
 import { MongooseTraineesRepository, MongooseUserRepository, MongooseGeographyRepository } from './repositories';
 import { GoogleOAuthService, TokenService, StorageService, UploadService, ImageService } from './services';
@@ -103,6 +104,7 @@ class Main {
     );
     const interactionController = new InteractionController(traineesRepository);
     const testController = new TestController(traineesRepository);
+    const strikeController = new StrikeController(traineesRepository);
     const searchController = new SearchController(traineesRepository);
     const geographyController = new GeographyController(geographyRepository);
     const dashboardController = new DashboardController();
@@ -118,6 +120,7 @@ class Main {
       interactionController,
       testController,
       profilePictureController,
+      strikeController,
       [authMiddleware]
     );
     const searchRouter = new SearchRouter(searchController, [authMiddleware]);
