+++ server/src/routes/TraineesRouter.ts
@@ -5,6 +5,7 @@ import {
   InteractionControllerType,
   TestControllerType,
   ProfilePictureControllerType,
+  StrikeControllerType,
 } from '../controllers';
 import Middleware from '../middlewares/Middleware';
 
@@ -13,6 +14,7 @@ export class TraineesRouter implements RouterType {
   private readonly interactionController: InteractionControllerType;
   private readonly profilePictureController: ProfilePictureControllerType;
   private readonly testController: TestControllerType;
+  private readonly strikeController: StrikeControllerType;
   private readonly middlewares: Middleware[];
 
   constructor(
@@ -20,12 +22,14 @@ export class TraineesRouter implements RouterType {
     interactionController: InteractionControllerType,
     testController: TestControllerType,
     profilePictureController: ProfilePictureControllerType,
+    strikeController: StrikeControllerType,
     middlewares: Middleware[] = []
   ) {
     this.traineeController = traineesController;
     this.interactionController = interactionController;
     this.testController = testController;
     this.profilePictureController = profilePictureController;
+    this.strikeController = strikeController;
     this.middlewares = middlewares;
   }
 
@@ -46,10 +50,10 @@ export class TraineesRouter implements RouterType {
       this.profilePictureController.deleteProfilePicture.bind(this.profilePictureController)
     );
 
-    router.get('/:id/strikes', this.traineeController.getStrikes.bind(this.traineeController));
-    router.post('/:id/strikes', this.traineeController.addStrike.bind(this.traineeController));
-    router.put('/:id/strikes/:strikeId', this.traineeController.updateStrike.bind(this.traineeController));
-    router.delete('/:id/strikes/:strikeId', this.traineeController.deleteStrike.bind(this.traineeController));
+    router.get('/:id/strikes', this.strikeController.getStrikes.bind(this.strikeController));
+    router.post('/:id/strikes', this.strikeController.addStrike.bind(this.strikeController));
+    router.put('/:id/strikes/:strikeId', this.strikeController.updateStrike.bind(this.strikeController));
+    router.delete('/:id/strikes/:strikeId', this.strikeController.deleteStrike.bind(this.strikeController));
 
     router.get('/:id/interactions', this.interactionController.getInteractions.bind(this.interactionController));
     router.post('/:id/interactions', this.interactionController.addInteraction.bind(this.interactionController));
